+++ MoltenVK/MoltenVK/GPUObjects/MVKDevice.h
@@ -357,9 +357,6 @@ public:
 		return _metalFeatures.argumentBuffers && getMVKConfig().useMetalArgumentBuffers != MVK_CONFIG_USE_METAL_ARGUMENT_BUFFERS_NEVER;
 	};
 
-	/** Returns whether native texture atomics are supported and should be used. */
-	bool useNativeTextureAtomics() { return _metalFeatures.nativeTextureAtomics; }
-
 	/** Returns the MTLStorageMode that matches the Vulkan memory property flags. */
 	MTLStorageMode getMTLStorageModeFromVkMemoryPropertyFlags(VkMemoryPropertyFlags vkFlags);